Adds support for voice input + difficulty limit
Voice input:
Added support for input using speech when on chrome.
It used the x-webkit-speech attribute, so it is limited to webkit.
If the number recognized is the answer, the answer will be accepted
right away. Else, if the number is not recognized or answer seems
wrong, It will wait for submit button to be clicked.

Difficulty limit:
Adds a variable maxNuber that dictates the maximum number that
will appear in any question. This variable will be used to implement
a dynamic difficulty switch in future.